+++ semantic_navigation_bt/include/semantic_navigation_bt/action/get_adjacent_regions_service.hpp
@@ -1,0 +1,77 @@
+// Copyright (c) 2026 Alberto J. Tudela Roldán
+// Copyright (c) 2026 Grupo Avispa, DTE, Universidad de Málaga
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef SEMANTIC_NAVIGATION_BT__ACTION__GET_ADJACENT_REGIONS_SERVICE_HPP_
+#define SEMANTIC_NAVIGATION_BT__ACTION__GET_ADJACENT_REGIONS_SERVICE_HPP_
+
+#include <memory>
+#include <string>
+#include <vector>
+
+#include "nav2_behavior_tree/bt_service_node.hpp"
+#include "semantic_navigation_msgs/srv/get_adjacent_regions.hpp"
+
+namespace semantic_navigation_bt
+{
+
+using nav2_behavior_tree::BtServiceNode;
+
+/**
+ * @brief A nav2_behavior_tree::BtServiceNode class that wraps
+ * semantic_navigation_msgs::srv::GetAdjacentRegions
+ */
+class GetAdjacentRegionsService
+  : public BtServiceNode<semantic_navigation_msgs::srv::GetAdjacentRegions>
+{
+public:
+  /**
+   * @brief A constructor for semantic_navigation_bt::GetAdjacentRegionsService
+   * @param service_node_name Service name this node creates a client for
+   * @param conf BT node configuration
+   */
+  GetAdjacentRegionsService(
+    const std::string & service_node_name, const BT::NodeConfiguration & conf);
+
+  /**
+   * @brief The main override required by a BT service to fill the request.
+   */
+  void on_tick() override;
+
+  /**
+   * @brief Override the on_completion method to set the output port with the adjacent regions
+   * after the completion of the service.
+   * @param response The response from the service
+   * @return BT::NodeStatus Returns SUCCESS if the adjacent regions are retrieved correctly
+   */
+  BT::NodeStatus on_completion(
+    std::shared_ptr<semantic_navigation_msgs::srv::GetAdjacentRegions::Response> response) override;
+
+  /**
+   * @brief Creates list of BT ports
+   * @return BT::PortsList Containing node-specific ports
+   */
+  static BT::PortsList providedPorts()
+  {
+    return providedBasicPorts(
+      {
+        BT::InputPort<std::string>("region_name", "Region name"),
+        BT::OutputPort<std::vector<std::string>>("adjacent_regions", "Adjacent regions")
+      });
+  }
+};
+
+}  // namespace semantic_navigation_bt
+
+#endif  // SEMANTIC_NAVIGATION_BT__ACTION__GET_ADJACENT_REGIONS_SERVICE_HPP_